% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_uarizona.R
\name{theme_uarizona}
\alias{theme_uarizona}
\title{Implements theme components.}
\usage{
theme_uarizona(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{the base size}

\item{base_family}{the base family}

\item{base_line_size}{the baseline size}

\item{base_rect_size}{the base rect}
}
\value{
a plot with UArizona colors
}
\description{
Defines the overall aesthetic and thematic features of the plot. This
function specifies simple background, grid line, text, and legend arguments
to create minimalist design. Its use is intended for ggplot objects.
}
\examples{

library(ggplot2)
library(palmerpenguins)

# default
p <- ggplot(penguins, aes(x = bill_depth_mm, y = bill_length_mm, color = species)) +
  geom_point() +
  labs(
    title = "Bill length and depth of penguins",
    subtitle = "Dimensions for Adelie, Chinstrap, and Gentoo Penguins",
    x = "Bill depth (mm)",
    y = "Bill length (mm)",
    color = "Species",
    caption = "Source: palmerpenguins package."
  )
p

# vs. with UArizona theme
p +
  theme_uarizona()

# vs. with UArizona theme and scale
p +
  scale_uarizona_color_discrete() +
  theme_uarizona()

# with UArizona theme, scale, and further customization to theme
p +
  scale_uarizona_color_discrete() +
  theme_uarizona() +
  theme(
    plot.title = element_text(color = "red", size = 20),
    plot.background = element_rect(fill = "pink", color = "yellow"),
    panel.grid = element_blank()
  )
}
