% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uarizona_pal.R
\name{uarizona_pal}
\alias{uarizona_pal}
\title{UArizona Color Palette (Discrete) and Scales}
\usage{
uarizona_pal()
}
\value{
Character vector of UArizona palette HEX codes.
}
\description{
An eight-color colorblind friendly qualitative discrete palette that is
based on colors on the UArizona branding guidelines.
}
\examples{
uarizona_pal()
}
\references{
https://brand.arizona.edu/applying-the-brand/colors
}
